requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-retry==2.8.3
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
bcrypt==4.2.0
blinker==1.8.2
boto3==1.35.40
botocore==1.35.40
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
decorator==4.4.2
distro==1.9.0
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.13.1
Flask==3.0.3
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.2.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.2
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.3
jiter==0.6.1
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
more-itertools==10.5.0
moviepy==1.0.3
mpmath==1.3.0
multidict==6.1.0
narwhals==1.9.3
networkx==3.2.1
nltk==3.9.1
numba==0.60.0
numpy==1.26.3
openai==1.51.2
openai-whisper @ git+https://github.com/openai/whisper.git@25639fc17ddc013d56c594bfbf7644f2185fad84
packaging==24.1
pandas==2.2.3
pillow==10.2.0
proglog==0.1.10
propcache==0.2.0
protobuf==5.28.2
psyc==0.1.0
psycopg2==2.9.9
pyarrow==17.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
# Removed PyJWT since it's related to JWT (which may be used for OTP scenarios)
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.9.2
rpds-py==0.20.0
s3transfer==0.10.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
semantic-version==2.10.0
setuptools==70.0.0
setuptools-rust==1.10.2
shap==0.46.0
six==1.16.0
slicer==0.0.8
smmap==5.0.1
sniffio==1.3.1
SpeechRecognition==3.10.4
streamlit==1.39.0
subprocess.run==0.0.8
sympy==1.12
tenacity==9.0.0
textblob==0.18.0.post0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.1
toml==0.10.2
torch==2.4.1+cu118
torchaudio==2.4.1+cu118
torchvision==0.19.1+cu118
tornado==6.4.1
tqdm==4.66.5
transformers==4.45.2
twilio==9.3.4
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==5.0.3
Werkzeug==3.0.4
whisper==1.1.10
yarl==1.15.4
